Extract FilterWithAnd
